fix:improve cursor tracking in browser input fields and refactored logic across multiple modules to improve snippet insertion, cursor tracking, and input handling
