+++ pages/content/src/types/index.ts
@@ -1,0 +1,12 @@
+export interface Snippet {
+  shortcut: string;
+  content: string;
+  name?: string;
+}
+
+export interface CursorInfo {
+  start: number;
+  end: number;
+  textBeforeCursor: string;
+  textAfterCursor: string;
+}
